Fix MongoDB client lifecycle in end-to-end tests

+++ api/e2e/books.e2e.js
@@ -17,48 +17,53 @@ const createApp = require("../src/app");
 describe("Test for books", () => {
   let app = null;
   let server = null;
+  let client = null;
   let database = null;
+
   beforeAll(async () => {
     app = createApp();
     server = app.listen(3001);
-    const client = new MongoClient(MONGO_URI, {
+    client = new MongoClient(MONGO_URI, {
       useNewUrlParser: true,
       useUnifiedTopology: true,
     });
     await client.connect();
     database = client.db(DB_NAME);
   }, 30000);
+
   afterAll(async () => {
     await server.close();
-    await database.dropDatabase();
+    if (database) {
+      await database.dropDatabase();
+    }
+    if (client) {
+      await client.close();
+    }
   });
+  test("should return a list books", async () => {
+    // Arrange
+    const seedData = await database.collection("books").insertMany([
+      {
+        name: "Book 1",
+        year: 1995,
+        author: "Sofia",
+      },
+      {
+        name: "Book 2",
+        year: 2000,
+        author: "Nicolas",
+      },
+    ]);
+    console.log(seedData);
 
-  describe("test for [GET] /api/v1/books", () => {
-    test("should return a list books", async () => {
-      // Arrange
-      const seedData = await database.collection("books").insertMany([
-        {
-          name: "Book 1",
-          year: 1995,
-          author: "Sofia",
-        },
-        {
-          name: "Book 2",
-          year: 2000,
-          author: "Nicolas",
-        },
-      ]);
-      console.log(seedData);
-
-      // Act
-      return request(app)
-        .get("/api/v1/books")
-        .expect(200)
-        .then(({ body }) => {
-          console.log(body);
-          // Assert
-          expect(body.length).toEqual(seedData.insertedCount);
-        });
-    });
+    // Act
+    return request(app)
+      .get("/api/v1/books")
+      .expect(200)
+      .then(({ body }) => {
+        console.log(body);
+        // Assert
+        expect(body.length).toEqual(seedData.insertedCount);
+      });
   });
 });
